requirements.txt
Django==4.0.6
python-gettext==4.0
django-extensions==3.1.5
isort==5.10.1
black==22.6.0
django-allauth==0.51.0
django-crispy-forms==1.14.0
factory_boy==3.2.1
pytest-cov==3.0.0